[fix] wrong type

+++ tests/mcd/conftest.py
@@ -29,6 +29,6 @@ def data(conf):
     dataset = TensorDataset(
         torch.tensor(user_id, dtype=torch.int64),
         torch.tensor(item_id, dtype=torch.int64),
-        torch.tensor(score, dtype=torch.int64)
+        torch.tensor(score, dtype=torch.float)
     )
     return DataLoader(dataset, batch_size=batch_size)
